+++ SharePointFramework/PortfolioWebParts/src/components/PortfolioAggregation/ColumnFormPanel/ColumnFormPanelFooter/types.ts
@@ -1,6 +1,6 @@
 import { useColumnFormPanel } from '../useColumnFormPanel'
 
-export type ColumnFormPanelFooterProps =Pick<
-    ReturnType<typeof useColumnFormPanel>,
-    'onSave' | 'onDeleteColumn' | 'isEditing' | 'isSaveDisabled' | 'isDeleteDisabled'
-  > 
+export type ColumnFormPanelFooterProps = Pick<
+  ReturnType<typeof useColumnFormPanel>,
+  'onSave' | 'onDeleteColumn' | 'isEditing' | 'isSaveDisabled' | 'isDeleteDisabled'
+>
